Match types for mixed precision processing
Former-commit-id: f65719cae61b28a6b535df9ae2cf8b3a2fe5570f

+++ tokun/evaluation.py
@@ -21,4 +21,4 @@ def intersection(left: str, right: str) -> float:
 # ROBUSTNESS ##################################################################
 
 def neighbors(point: tf.Tensor, radius: float, count: int) -> tf.Tensor:
-    return point + tf.random.uniform(shape=(count, point.shape[-1]), minval=-radius, maxval=radius, dtype=tf.dtypes.float32)
+    return point + tf.random.uniform(shape=(count, point.shape[-1]), minval=-radius, maxval=radius, dtype=point.dtype)
